2022/17: Use a cache for loop detection in part 2.
My original implementation didn't work with the example
input, and I think I also forgot to track the current jet of
gas. I guess I was just lucky that it worked with my input.
Switch to a more generalized solution that keeps a cache of
seen states (including bitfields describing the top 40 rows
of the chamber).

This is imperfect (there could be a gap going down more than
40 rows for some inputs) and there's also an off-by-one bug
somewhere, but it's better than what I had before, I guess.